moved sortedIntersectionCount to java Hotspots class and a bit of cleanup
