--- core/src/main/java/com/klibisz/elastiknn/JavaDummy.java
@@ -1,6 +1,0 @@
-package com.klibisz.elastiknn;
-
-/**
- * Gradle complains if there's no Java source code.
- */
-class JavaDummy { }
